Create README.md
I'm not entirely sure how to write this yet. I just want to explain the kickstart files for now. There will be a base kickstart file that just has partitioning, NIC configuration and a boot path. The basic file is the one we will be using to start. Other kickstart files are for experimentation and testing for now.